implémentation/chatBot/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
dataclasses-json==0.5.14
effdet==0.4.1
emoji==2.8.0
faiss-cpu==1.7.4
filelock==3.12.3
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
greenlet==2.0.2
huggingface-hub==0.17.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langchain==0.0.286
langsmith==0.0.35
layoutparser==0.3.4
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==10.0.0
portalocker==2.7.0
protobuf==4.24.3
pycocotools==2.0.7
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.0.9
pypdfium2==4.20.0
pyreadline3==3.4.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scipy==1.11.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.4.0
timm==0.9.7
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unstructured==0.10.14
unstructured-inference==0.5.25
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.16.2